Add virtual inheritance.
With this, we add cross-relations between the various base classes
defined in this package. Now, every OptGoalEnv now also is
an OptEnv; every SeparableOptGoalEnv also is an OptGoalEnv; etc.

The only exception is the link SeparableEnv <-> SeparableGoalEnv (and
the optimizable variants). The reason is that both classes define
`compute_reward()` methods that are compatible in their signature, but
not in their meanings. Thus, it cannot be said in good conscience that
one is a subclass of the other.