Unmark Problem.render() as an abstract method.
This allows subclasses to skip implementing this method if they don't
support rendering (`metadata['render.modes'] == []`). Before, they
would've had to write a shim that simply calls the super method.

Note that Problem.render() still raises a `NotImplementedError` if
called. This is for compatibility with subclasses that call it on
unknown render modes.